Match msg_get_glyph
And fix the type definition for draw_msg, that's it really.

+++ include/functions.h
@@ -505,7 +505,7 @@ void mdl_draw_hidden_panel_surface(Gfx** arg0, u16 treeIndex);
 s32 func_8011CFBC(void);
 s32 set_screen_overlay_center_worldpos(void);
 s32 mdl_get_next_texture_address(s32);
-void draw_msg(s32 msgID, s32 posX, s32 posY, s32 opacity, s32 palette, s32 style);
+void draw_msg(s32 msgID, s32 posX, s32 posY, s32 opacity, s32 palette, u8 style);
 void get_background_color_blend(u8* r, u8* g, u8* b, u8* a);
 
 s8 set_global_byte(s32 index, s32 value);
